requirements.txt
aiogram==3.15.0
httpx==0.28.1
aiosqlite==0.20.0
python-dotenv==1.0.1
openai==1.61.1
selectolax==0.4.6
